requirements.txt
apiclient==1.0.4
build==0.7.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
DateTime==4.4
discovery==0.0.4
gaapi4py==1.2.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
gspread==5.3.2
httplib2==0.20.4
idna==3.3
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pep517==0.12.0
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
xyaml==1.0.2
zope.interface==5.4.0
